Add support for setting IPV6_V6ONLY option
On some operating systems, opening a socket and having it listen on the
IPv6 unspecified address (::) will silently open a dual-stack socket,
listening both for IPv4 and IPv6 connections. On Linux, this is often
the default, although this behavior can be controlled by the
sys.net.ipv6.bindv6only sysctl.

However, some applications may want to ensure that IPv6 sockets always
listen to IPv6 connections *only*, e.g. for portability reasons. For
example, a server program may ship with a configuration example that has
it listen on IPv4 and IPv6 on all interfaces. According to whether IPv6
sockets are single- or dual-stack, this configuration example may fail
on some hosts. The only way to do this portably is to set the
IPV6_V6ONLY socket option to 1 with setsockopt(2) right after creating
the socket and creating one socket per family in the application.

This commit adds the ipv6_v6only option to procket:open/2 and the
corresponding -O command-line option in the procket binary. If left
unspecified, procket will use the OS environment’s default behavior
regarding dual-stack IPv6 sockets. If explicitly set to true or false,
procket will call setsockopt(2) to set IPV6_V6ONLY to 1 or 0
respectively.

See also: ipv6(7) on Linux or ip6(4) on FreeBSD.